refactor(exchange-oracle): split job_creation into per-task package
Break the 3k-line src/handlers/job_creation.py into a package with one
file per task builder:
- builders/vision/{base,basic,boxes_from_points,skeletons_from_boxes}.py
- builders/audio/transcription.py (stub, raises NotImplementedError)
- factory.py (create_task), exceptions.py, utils.py
__init__.py exports only create_task.

Move DM_DATASET_FORMAT_MAPPING / DM_GT_DATASET_FORMAT_MAPPING to a new
src/core/tasks/cvat_formats.py and repoint job_export.py and the webhook
test's mock-patch paths to the new module locations.
