+++ packages/examples/cvat/exchange-oracle/src/handlers/job_creation/__init__.py
@@ -1,0 +1,3 @@
+from src.handlers.job_creation.factory import create_task
+
+__all__ = ["create_task"]
