+++ packages/examples/cvat/exchange-oracle/src/handlers/job_creation/builders/audio/transcription.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from src.handlers.job_creation.builders.vision.base import _TaskBuilderBase
+
+
+class AudioTranscriptionTaskBuilder(_TaskBuilderBase):
+    """
+    Handles task creation for the AUDIO_TRANSCRIPTION task type.
+
+    Stub: audio transcription task creation is not implemented yet.
+    """
+
+    def build(self) -> None:
+        raise NotImplementedError("Audio transcription task creation is not implemented yet")
