+++ packages/examples/cvat/exchange-oracle/src/handlers/job_creation/exceptions.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+
+class DatasetValidationError(Exception):
+    pass
+
+
+class MismatchingAnnotations(DatasetValidationError):
+    pass
+
+
+class TooFewSamples(DatasetValidationError):
+    pass
+
+
+class InvalidCategories(DatasetValidationError):
+    pass
+
+
+class InvalidImageInfo(DatasetValidationError):
+    pass
+
+
+class InvalidCoordinates(DatasetValidationError):
+    pass
+
+
+class InvisibleSkeletonError(DatasetValidationError):
+    pass
